# openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Management System API
  description: Centralized API documentation for the Restaurant Management System. Access and test endpoints across all 8 microservices.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: API Gateway (Main Entrypoint)
  - url: http://localhost:5001
    description: User Service
  - url: http://localhost:5002
    description: Menu Service
  - url: http://localhost:5003
    description: Inventory Service
  - url: http://localhost:5004
    description: Order Service
  - url: http://localhost:5005
    description: Payment Service
  - url: http://localhost:5006
    description: Table Service
  - url: http://localhost:5007
    description: Branch Service
  - url: http://localhost:5008
    description: Report Service
paths:
...
  /api/orders/:
    post:
      tags:
        - Orders
      summary: Create a new order
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: X-Verification-Token
          schema:
            type: string
          required: true
          description: Verification token obtained from OTP verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - branch_id
                - table_id
                - type
              properties:
                branch_id:
                  type: integer
                table_id:
                  type: integer
                type:
                  type: string
                  enum: [DINE_IN, TAKE_AWAY, DELIVERY]
                notes:
                  type: string
      responses:
        '201':
          description: Order created successfully
    get:
      tags:
        - Orders
      summary: List all orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of orders
...
  /api/orders/{order_id}/items:
    post:
      tags:
        - Order Items
      summary: Add items to an order
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - menu_item_id
                - quantity
              properties:
                menu_item_id:
                  type: integer
                quantity:
                  type: integer
                notes:
                  type: string
      responses:
        '200':
          description: Items added to order
  /api/orders/{order_id}/items/{item_id}:
    put:
      tags:
        - Order Items
      summary: Update order item quantity/notes
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                notes:
                  type: string
      responses:
        '200':
          description: Order item updated
    delete:
      tags:
        - Order Items
      summary: Remove item from order
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item removed from order
  /api/orders/{order_id}/items/{item_id}/cancel:
    put:
      tags:
        - Order Items
      summary: Cancel a preparation item in kitchen (Manager/Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item cancelled in kitchen
  /api/orders/{order_id}/send-kitchen:
    post:
      tags:
        - Order Workflow
      summary: Send order items to the kitchen
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order items sent to kitchen
  /api/orders/{order_id}/preparing:
    put:
      tags:
        - Order Workflow
      summary: Mark order status as PREPARING (Kitchen only)
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Status updated to preparing
  /api/orders/{order_id}/done:
    put:
      tags:
        - Order Workflow
      summary: Mark order status as DONE/Ready (Kitchen only)
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Status updated to ready
  /api/orders/{order_id}/receive:
    put:
      tags:
        - Order Workflow
      summary: "Customer receives the food (status: COMPLETED)"
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Status updated to completed
  /api/orders/{order_id}/cancel:
    put:
      tags:
        - Order Workflow
      summary: Cancel the order
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order cancelled
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT